Refactor buttons; drop more pcss
