+++ src/modals/AccessingProjectModal.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import { Buttons } from '~/shared/components/Buttons'
+import { Buttons } from '~/components/Buttons'
 import PrestyledLoadingIndicator from '~/shared/components/LoadingIndicator'
 import { COLORS } from '~/shared/utils/styled'
 import Modal from './Modal'
