+++ src/modals/ConfirmPurchaseModal.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import { Buttons } from '~/shared/components/Buttons'
+import { Buttons } from '~/components/Buttons'
 import Modal from './Modal'
 import { Footer } from './BaseModal'
