--- src/shared/components/Buttons/buttons.pcss
@@ -1,25 +1,0 @@
-.buttons {
-  display: flex;
-  justify-content: flex-end;
-  width: 100%;
-
-  /* prevents border rendering glitches in chrome */
-  transform: translate3d(0, 0, 0);
-}
-
-.buttons:empty {
-  display: none;
-}
-
-.buttons > * {
-  padding-right: 15px;
-  padding-left: 15px;
-}
-
-.buttons > :not(.link) {
-  min-width: 7rem;
-}
-
-.buttons > * + * {
-  margin-left: 1rem;
-}
